+++ packages/observe-sequence/package.js
@@ -1,14 +1,14 @@
 Package.describe({
   summary: "Observe changes to various sequence types such as arrays, cursors and objects",
-  version: "1.0.20"
+  version: '2.0.0-alpha300.17',
 });
 
 Package.onUse(function (api) {
-  api.use('tracker@1.2.0');
+  api.use('tracker@1.3.2');
   api.use('mongo-id@1.0.8');  // for idStringify
-  api.use('diff-sequence@1.1.1');
-  api.use('random@1.2.0');
-  api.use('ecmascript@0.13.2');
+  api.use('diff-sequence@1.1.2');
+  api.use('random@1.2.1');
+  api.use('ecmascript@0.16.8-alpha300.17');
   api.export('ObserveSequence');
   api.addFiles(['observe_sequence.js']);
 });