+++ packages/spacebars-compiler/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: 'spacebars-compiler',
   summary: "Compiler for Spacebars template language",
-  version: '1.3.1',
+  version: '2.0.0-alpha300.17',
   git: 'https://github.com/meteor/blaze.git'
 });
 
@@ -10,11 +10,11 @@ Npm.depends({
 });
 
 Package.onUse(function (api) {
-  api.use('ecmascript@0.15.1');
+  api.use('ecmascript@0.16.8-alpha300.17');
 
-  api.use('htmljs@1.1.1');
-  api.use('html-tools@1.1.3');
-  api.use('blaze-tools@1.1.3');
+  api.use('htmljs@2.0.0-alpha300.17');
+  api.use('html-tools@2.0.0-alpha300.17');
+  api.use('blaze-tools@2.0.0-alpha300.17');
 
   api.export('SpacebarsCompiler');